Create ephem.py
For the course assignment, we received code for a MATLAB function from the course’s TA, Dylan Clay, that directly loaded in the data we needed from the ephemeris kernel necessary. In order to get this data for my project, I used Python's "jplephem" module to access the Satellite Planet Kernel (SPK) files that NASA JPL provides. These are used in conjunction with their SPICE toolkit. Once I had done this and downloaded a .bsp file that gave ephemeris data from 1900 to 2050, I implemented this data into my code.